+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Add possibility to display archived sources on active-sources page
 - Update the way active offers are filtered to match only given subsidiary
 - Add webhook to add new process from cognito form
+- Add others process on process detail for a candidate
 
 # v1.16.0 (2022-11-14)
 - More refined factories